Renaming UserDetails base component and its usages
