+++ frontend/awx/access/users/UserPage/AwxUserDetails.cy.tsx
@@ -1,4 +1,4 @@
-import { UserDetails } from './UserDetails';
+import { AwxUserDetails } from './AwxUserDetails';
 import mockAwxUser from '../../../../../cypress/fixtures/awxUser.json';
 import { formatDateString } from '../../../../../framework/utils/formatDateString';
 
@@ -6,7 +6,7 @@ describe('User details', () => {
   it('Renders first & last name, username, email, orgs, last login, auth type, created & modified timestamps', () => {
     cy.intercept('/api/v2/users/*', { fixture: 'awxUser.json' });
     cy.intercept('/api/v2/users/*/organizations/', { fixture: 'organizations.json' });
-    cy.mount(<UserDetails />);
+    cy.mount(<AwxUserDetails />);
     cy.get('[data-cy="first-name"]').should('have.text', 'Org');
     cy.get('[data-cy="last-name"]').should('have.text', 'Admin');
     cy.get('[data-cy="email"]').should('have.text', 'firstname@lastname.com');
