+++ frontend/eda/UserAccess/Users/UserPage/EdaUserDetails.tsx
@@ -1,13 +1,13 @@
 /* eslint-disable react/prop-types */
 import { useParams } from 'react-router-dom';
 import { LoadingPage } from '../../../../../framework/components/LoadingPage';
-import { UserDetailsBase, UserDetailsType } from '../../../../common/access/UserDetailsBase';
+import { UserDetails, UserDetailsType } from '../../../../common/access/UserDetails';
 import { useGet } from '../../../../common/crud/useGet';
 import { SWR_REFRESH_INTERVAL } from '../../../constants';
 import { EdaUser } from '../../../interfaces/EdaUser';
 import { edaAPI } from '../../../api/eda-utils';
 
-export function UserDetails() {
+export function EdaUserDetails() {
   const params = useParams<{ id: string }>();
   const { data: user } = useGet<EdaUser>(edaAPI`/users/${params.id ?? ''}/`, undefined, {
     refreshInterval: SWR_REFRESH_INTERVAL,
@@ -16,7 +16,7 @@ export function UserDetails() {
 
   return (
     <>
-      <UserDetailsBase user={user as UserDetailsType} />
+      <UserDetails user={user as UserDetailsType} />
     </>
   );
 }
