+++ frontend/eda/UserAccess/Users/UserPage/MyDetails.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/prop-types */
 import { LoadingPage } from '../../../../../framework/components/LoadingPage';
-import { UserDetailsBase, UserDetailsType } from '../../../../common/access/UserDetailsBase';
+import { UserDetails, UserDetailsType } from '../../../../common/access/UserDetails';
 import { useGet } from '../../../../common/crud/useGet';
 import { edaAPI } from '../../../api/eda-utils';
 import { EdaUser } from '../../../interfaces/EdaUser';
@@ -12,7 +12,7 @@ export function MyDetails() {
 
   return (
     <>
-      <UserDetailsBase user={user as UserDetailsType} />
+      <UserDetails user={user as UserDetailsType} />
     </>
   );
 }
