+++ frontend/eda/useEdaNavigation.tsx
@@ -42,7 +42,7 @@ import { ControllerTokens } from './UserAccess/Users/UserPage/ControllerTokens';
 import { UserPage } from './UserAccess/Users/UserPage/UserPage';
 import { MyPage } from './UserAccess/Users/UserPage/MyPage';
 import { MyDetails } from './UserAccess/Users/UserPage/MyDetails';
-import { UserDetails } from './UserAccess/Users/UserPage/UserDetails';
+import { EdaUserDetails } from './UserAccess/Users/UserPage/EdaUserDetails';
 import { RuleAuditPage } from './views/RuleAudit/RuleAuditPage/RuleAuditPage';
 import { RuleAuditActions } from './views/RuleAudit/RuleAuditPage/RuleAuditActions';
 import { RuleAuditEvents } from './views/RuleAudit/RuleAuditPage/RuleAuditEvents';
@@ -327,7 +327,7 @@ export function useEdaNavigation() {
                       {
                         id: EdaRoute.UserDetails,
                         path: 'details',
-                        element: <UserDetails />,
+                        element: <EdaUserDetails />,
                       },
                       {
                         id: EdaRoute.UserTokens,
